refactor(desktop): remove legacy check-for-user-changes tool and relocate executor response types
