+++ src/tools/desktop/tools.ts
@@ -54,7 +54,6 @@ import { getLookupWorkbookSchemaTool } from './search/lookupWorkbookSchema.js';
 import { getSearchCommandsTool } from './search/searchCommands.js';
 import { getSearchExamplesTool } from './search/searchExamples.js';
 import { getSearchWorkbookExamplesTool } from './search/searchWorkbookExamples.js';
-import { getCheckForUserChangesTool } from './session/checkForUserChanges.js';
 import { getListInstancesTool } from './session/listInstances.js';
 import { getInjectTemplateTool } from './template/injectTemplate.js';
 import { getActivateSheetTool } from './workbook/activateSheet.js';
@@ -69,7 +68,6 @@ import { getRefineWorksheetTool } from './worksheet/refineWorksheet.js';
 
 export const desktopToolFactories = [
   getListInstancesTool,
-  getCheckForUserChangesTool,
   getGetWorkbookXmlTool,
   getApplyWorkbookTool,
   getActivateSheetTool,
